Move time funcs to util
