+++ go.mod
@@ -8,11 +8,10 @@ require (
 	github.com/golang/mock v1.6.0 // indirect
 	github.com/jinzhu/now v1.1.5 // indirect
 	github.com/joho/godotenv v1.3.0
-	github.com/stretchr/testify v1.8.0 // indirect
-	golang.org/x/net v0.0.0-20210614182718-04defd469f4e
+	golang.org/x/net v0.10.0
 	golang.org/x/oauth2 v0.0.0-20210514164344-f6687ab2804c
-	golang.org/x/term v0.0.0-20220411215600-e5f449aeb171 // indirect
 	google.golang.org/api v0.48.0
 	gorm.io/driver/mysql v1.3.3
-	gorm.io/gorm v1.23.5
+	gorm.io/driver/postgres v1.5.4 // indirect
+	gorm.io/gorm v1.25.5
 )
